fix(witype): Relocated work item type model
