+++ src/app/work-item/work-item-detail/work-item-detail.component.spec.ts
@@ -40,7 +40,7 @@ import { IterationModel } from '../../models/iteration.model';
 import { IterationService } from '../../iteration/iteration.service';
 import { LinkType } from '../../models/link-type';
 import { WorkItem } from '../../models/work-item';
-import { WorkItemType } from '../work-item-type';
+import { WorkItemType } from '../../models/work-item-type';
 import { WorkItemService } from '../work-item.service';
 
 import { WorkItemLinkComponent } from './work-item-link/work-item-link.component';
